requirements.txt
asgiref==3.2.5
astroid==2.3.3
attrs==19.3.0
awacs==0.9.8
awscli==1.18.26
awsebcli==3.17.1
bcrypt==3.1.7
blessed==1.17.3
boto3==1.12.26
botocore==1.15.26
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
cfn-flip==1.2.2
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==2.8
Django==3.0.4
djangorestframework==3.11.0
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
importlib-metadata==1.5.0
isort==4.3.21
jmespath==0.9.5
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
paramiko==2.7.1
pathspec==0.5.9
pep8==1.7.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycparser==2.20
pylint==2.4.4
PyNaCl==1.3.0
pyrsistent==0.15.7
python-dateutil==2.8.0
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
requests==2.20.1
rsa==3.4.2
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
troposphere==2.6.0
typed-ast==1.4.1
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.57.0
wrapt==1.11.2
zipp==3.1.0
